# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Gestion Comptes API",
        "description": "API pour la gestion des comptes bancaires",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur local de développement"
        },
        {
            "url": "https://gestion-comptes-bancaire-marra-ousmane.onrender.com",
            "description": "Serveur de production"
        }
    ],
    "paths": {
...
        "/api/v1/clients/{compteId}": {
            "patch": {
                "tags": [
                    "Clients"
                ],
                "summary": "Mettre à jour les informations d'un client via son compte",
                "description": "Met à jour partiellement les informations d'un client en utilisant l'ID d'un de ses comptes. Les modifications sont appliquées dans la base Render.",
                "operationId": "updateClientViaCompte",
                "parameters": [
                    {
                        "name": "compteId",
                        "in": "path",
                        "description": "ID du compte appartenant au client",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "prenom": {
                                        "type": "string",
                                        "example": "John",
                                        "nullable": true
                                    },
                                    "nom": {
                                        "type": "string",
                                        "example": "Doe",
                                        "nullable": true
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john.doe@example.com",
                                        "nullable": true
                                    },
                                    "telephone": {
                                        "type": "string",
                                        "example": "+221771234567",
                                        "nullable": true
                                    },
                                    "adresse": {
                                        "type": "string",
                                        "example": "123 Main St",
                                        "nullable": true
                                    },
                                    "nci": {
                                        "type": "string",
                                        "example": "1234567890123",
                                        "nullable": true
                                    },
                                    "date_naissance": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1990-01-01",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Client mis à jour avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "string",
                                                    "example": "uuid"
                                                },
                                                "prenom": {
                                                    "type": "string",
                                                    "example": "John",
                                                    "nullable": true
                                                },
                                                "nom": {
                                                    "type": "string",
                                                    "example": "Doe"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "john.doe@example.com"
                                                },
                                                "telephone": {
                                                    "type": "string",
                                                    "example": "+221771234567"
                                                },
                                                "adresse": {
                                                    "type": "string",
                                                    "example": "123 Main St",
                                                    "nullable": true
                                                },
                                                "nci": {
                                                    "type": "string",
                                                    "example": "1234567890123",
                                                    "nullable": true
                                                },
                                                "date_naissance": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "1990-01-01",
                                                    "nullable": true
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Client mis à jour avec succès"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Client non trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Client non trouvé"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Erreur de validation"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/clients/telephone/{telephone}": {
            "get": {
                "tags": [
                    "Clients"
                ],
                "summary": "Obtenir les détails d'un client par téléphone",
                "description": "Récupère les informations détaillées d'un client en utilisant son numéro de téléphone",
                "operationId": "getClientByTelephone",
                "parameters": [
                    {
                        "name": "telephone",
                        "in": "path",
                        "description": "Numéro de téléphone du client",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "+221771234567"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails du client récupérés",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "string",
                                                    "example": "uuid"
                                                },
                                                "prenom": {
                                                    "type": "string",
                                                    "example": "John",
                                                    "nullable": true
                                                },
                                                "nom": {
                                                    "type": "string",
                                                    "example": "Doe"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "john.doe@example.com"
                                                },
                                                "telephone": {
                                                    "type": "string",
                                                    "example": "+221771234567"
                                                },
                                                "nci": {
                                                    "type": "string",
                                                    "example": "1234567890123",
                                                    "nullable": true
                                                },
                                                "client": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "string",
                                                            "example": "uuid"
                                                        },
                                                        "adresse": {
                                                            "type": "string",
                                                            "example": "123 Main St",
                                                            "nullable": true
                                                        },
                                                        "date_naissance": {
                                                            "type": "string",
                                                            "format": "date",
                                                            "example": "1990-01-01",
                                                            "nullable": true
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Détails du client"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Client non trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Client non trouvé"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "JWT Authorization header using the Bearer scheme. Example: \"Authorization: Bearer {token}\"",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ]
}